feat:api integration

+++ src/main/java/com/example/demo/repo/PlayerRepo.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.example.demo.model.Players;
 
-public interface PlayerRepo extends JpaRepository<Players,Integer> {
+public interface PlayerRepo extends JpaRepository<Players,Long> {
 
 }
