+++ src/main/java/com/example/demo/service/PlyerService.java
@@ -4,12 +4,14 @@ import java.util.List;
 
 import com.example.demo.model.Players;
 
+
+
 public interface PlyerService {
 
     public void savePlayer(Players players);
 
     public List<Players>getAllPlayers();
 
-    public Players  getPlayerById(int playerId);
+    public Players  getPlayerById(Long playerId);
 
 }
